feat: 敏感信息存储支持国密 #2055
优化代码结构
