+++ src/backend/commons/common-jwt/src/main/java/com/tencent/bk/job/common/jwt/JwtManager.java
@@ -1,4 +1,4 @@
-package com.tencent.bk.job.common.util.jwt;
+package com.tencent.bk.job.common.jwt;
 
 public interface JwtManager {
     String getToken();
