+++ src/backend/commons/common-security/src/main/java/com/tencent/bk/job/common/security/jwt/AutoUpdateJwtManager.java
@@ -1,16 +1,12 @@
 package com.tencent.bk.job.common.security.jwt;
 
-import com.tencent.bk.job.common.util.jwt.BasicJwtManager;
+import com.tencent.bk.job.common.jwt.BasicJwtManager;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 
-import java.io.IOException;
-import java.security.GeneralSecurityException;
-
 @Slf4j
 public class AutoUpdateJwtManager extends BasicJwtManager {
-    public AutoUpdateJwtManager(String privateKeyBase64, String publicKeyBase64)
-        throws IOException, GeneralSecurityException {
+    public AutoUpdateJwtManager(String privateKeyBase64, String publicKeyBase64) {
         super(privateKeyBase64, publicKeyBase64);
     }
