+++ src/backend/commons/common-web/src/main/java/com/tencent/bk/job/common/web/interceptor/ServiceSecurityInterceptor.java
@@ -24,7 +24,7 @@
 
 package com.tencent.bk.job.common.web.interceptor;
 
-import com.tencent.bk.job.common.util.jwt.JwtManager;
+import com.tencent.bk.job.common.jwt.JwtManager;
 import com.tencent.bk.job.common.web.exception.ServiceNoAuthException;
 import com.tencent.bk.job.common.web.util.ProfileUtil;
 import lombok.extern.slf4j.Slf4j;
