+++ src/backend/settings.gradle
@@ -33,6 +33,7 @@ include 'commons:cmdb-sdk-ext'
 include 'commons:paas-sdk'
 include 'commons:gse-sdk'
 include 'commons:common-iam'
+include 'commons:common-jwt'
 include 'commons:common-utils'
 include 'commons:common-otel'
 include 'commons:common-redis'
